interface/requirements.txt
kafka-python==2.2.6
pandas==2.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
streamlit==1.45.0
